sm1-de-assignment-1/training/requirements.txt
pandas
google-cloud-storage
scikit-learn
flask
